Created a simulation for an airline
This simulation runs for 6000 seconds and simulates an airline with only one runway. Planes landing are given higher priority compared to planes taking off for this runway. At the end of the simulation, a summary of statistics is generated for information such as the number of planes landed, departed, crashed, average time for planes to land and take off.