Trees
Implemented a class for a splay tree data structure. When the program starts, a splay tree is built and printed by reading in values from a file named "in.dat". After this, an interactive mode begins and displays commands that allow you to search, insert delete, or splay. After each operation, a message will be printed and the new tree will be displayed.