Simulations
The purpose of this simulation is to compare the number of comparisons of each sorting algorithm and determine which algorithm performed the fastest. Every thousand random numbers generated, a table will print with the number of comparisons each algorithm made. Sorting algorithms such as quicksort, heapsort, mergesort, shell sort, and insertion sort were tested.